feat[character]: Add attributes
This commit alone compiles, but the tests won't pass. This is
because a) there's no good mechanism yet to add/change attributes,
b) because the JSON structure of the character files is too complicated
for the structure of data once serialized with serde.

I'm reworking the JSONs in a separate commit (commits will have to be
squashed later, or we won't have proper atomic commits).